tidy: misc small mechanical fixes (24 across 7 checks)
clang-tidy --fix for:
  modernize-deprecated-headers (4)   <stdint.h> -> <cstdint> et al.
  modernize-use-nullptr (2)
  modernize-use-using (1)
  modernize-use-nodiscard (2)
  readability-redundant-parentheses (4)
  readability-redundant-control-flow (1)
  readability-redundant-casting (3)

Build + 256/256 tests green.

+++ rosetta_core/src/TranslatorHelpers.cpp
@@ -644,7 +644,7 @@ auto translate_gpr(TranslationResult* result, int is_64bit, uint8_t reg, unsigne
     assert((reg & 0xFF) < 0x50 && "translate_gpr: called on non-GPR register");
 
     const int index = (int)(reg & 0xF);
-    const int size_class = (int)((reg >> 4) & 0xF);
+    const int size_class = ((reg >> 4) & 0xF);
 
     // -------------------------------------------------------------------------
     // Branch to BFX block (loc_20E9C) when:
